refactor(permission): split sandbox/mod.rs into submodules
Split sandbox/mod.rs (538 lines) into three files to comply with
code size limits:
- security.rs (~109 lines): SecurityPolicy + platform-specific stubs
- ipc.rs (~237 lines): IpcChannel, SandboxRequest/Response, IPC protocol
- mod.rs (~334 lines): Sandbox lifecycle, SandboxState, SandboxError

Also refactored Sandbox::spawn() into spawn + spawn_child_process +
confirm_engine_ready to keep function bodies under 50 lines.
Split impl Sandbox into 3 blocks (25/84/74 lines) under 100-line limit.

Added unit tests for SecurityPolicy, IPC serialization, and
SandboxState/SandboxError.

Source: issue #222
Type: refactor
